### What changes were proposed in this pull request?  When we calculate the quantile information from the peak executor metrics values for the distribution, there is a possibility of running into an `ArrayIndexOutOfBounds` exception when the metric values are empty. This PR addresses that and fixes it by returning an empty array if the values are empty.
 ### Why are the changes needed?
 Without these changes, when the withDetails query parameter is used to query the stages REST API, we encounter a partial JSON response since the peak executor metrics distribution cannot be serialized due to the above index error.

 ### Does this PR introduce _any_ user-facing change?
 No

 ### How was this patch tested?
 Added a unit test to test this behavior
