+++ core/src/main/scala/org/apache/spark/status/AppStatusStore.scala
@@ -31,6 +31,7 @@ import org.apache.spark.status.api.v1
 import org.apache.spark.storage.FallbackStorage.FALLBACK_BLOCK_MANAGER_ID
 import org.apache.spark.ui.scope._
 import org.apache.spark.util.Utils
+import org.apache.spark.util.Utils.getQuantilesValue
 import org.apache.spark.util.kvstore.KVStore
 
 /**
@@ -768,14 +769,6 @@ private[spark] class AppStatusStore(
             values.flatMap(_.peakMemoryMetrics))
       ))
     }
-  }
-
-  def getQuantilesValue(
-    values: IndexedSeq[Double],
-    quantiles: Array[Double]): IndexedSeq[Double] = {
-    val count = values.size
-    val indices = quantiles.map { q => math.min((q * count).toLong, count - 1) }
-    indices.map(i => values(i.toInt)).toIndexedSeq
   }
 
   def rdd(rddId: Int): v1.RDDStorageInfo = {
