+++ core/src/main/scala/org/apache/spark/util/Utils.scala
@@ -199,6 +199,18 @@ private[spark] object Utils extends Logging with SparkClassUtils {
     Try { classForName(clazz, initialize = false) }.isSuccess
   }
 
+  def getQuantilesValue(
+    values: IndexedSeq[Double],
+    quantiles: Array[Double]): IndexedSeq[Double] = {
+    val count = values.size
+    val indices = quantiles.map { q => math.min((q * count).toLong, count - 1) }
+    if (count > 0) {
+      indices.map(i => values(i.toInt)).toIndexedSeq
+    } else {
+      quantiles.map(_ => 0.0)
+    }
+  }
+
   /**
    * Run a segment of code using a different context class loader in the current thread
    */
